chore(plugins): move unrelated code from logging.rs into new, more tightly scoped, files
